+++ src/app/globals.css
@@ -1,5 +1,4 @@
 @import "tailwindcss";
-@tailwind utilities;
 
 
 :root {
