+++ src/app/seasons/[id]/page.tsx
@@ -1,6 +1,3 @@
-import RootLayout from "@/app/layout";
-import Link from "next/link";
-
 import SeasonConstellations from "./SeasonConstellations";
 
 /** それぞれの季節の星座を説明するページ */
@@ -8,10 +5,8 @@ export default function SeasonConstellationsDetailPage({
 	params,
 }: { params: { id: string } }) {
 	return (
-		<RootLayout>
-			<main className="max-w-screen-md mx-auto p-6 ">
-				<SeasonConstellations id={params.id} />
-			</main>
-		</RootLayout>
+		<main className="max-w-screen-md mx-auto p-6 ">
+			<SeasonConstellations id={params.id} />
+		</main>
 	);
 }
